Initial work for file format writer API
