+++ tests/io/test_fileformat.py
@@ -1,0 +1,24 @@
+# Licensed to the Apache Software Foundation (ASF) under one
+# or more contributor license agreements.  See the NOTICE file
+# distributed with this work for additional information
+# regarding copyright ownership.  The ASF licenses this file
+# to you under the Apache License, Version 2.0 (the
+# "License"); you may not use this file except in compliance
+# with the License.  You may obtain a copy of the License at
+#
+#   http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing,
+# software distributed under the License is distributed on an
+# "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+# KIND, either express or implied.  See the License for the
+# specific language governing permissions and limitations
+# under the License.
+
+
+def test_backward_compat_import() -> None:
+    """DataFileStatistics can still be imported from pyiceberg.io.pyarrow."""
+    from pyiceberg.io.fileformat import DataFileStatistics as dFS  # noqa: F401
+    from pyiceberg.io.pyarrow import DataFileStatistics  # noqa: F401
+
+    assert DataFileStatistics is dFS
